backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic
pydantic-settings
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
